Split up responsibilities of Rail class
