+++ src/pocketmine/block/DetectorRail.php
@@ -23,16 +23,12 @@ declare(strict_types=1);
 
 namespace pocketmine\block;
 
-class DetectorRail extends Rail{
+class DetectorRail extends RedstoneRail{
 
 	protected $id = self::DETECTOR_RAIL;
 
 	public function getName() : string{
 		return "Detector Rail";
-	}
-
-	public function canCurve() : bool{
-		return false;
 	}
 
 	//TODO
