+++ src/pocketmine/block/PoweredRail.php
@@ -23,14 +23,10 @@ declare(strict_types=1);
 
 namespace pocketmine\block;
 
-class PoweredRail extends Rail{
+class PoweredRail extends RedstoneRail{
 	protected $id = self::POWERED_RAIL;
 
 	public function getName() : string{
 		return "Powered Rail";
-	}
-
-	public function canCurve() : bool{
-		return false;
 	}
 }
