+++ src/pocketmine/block/RedstoneRail.php
@@ -23,13 +23,10 @@ declare(strict_types=1);
 
 namespace pocketmine\block;
 
-class ActivatorRail extends RedstoneRail{
+class RedstoneRail extends BaseRail{
+	protected const FLAG_POWERED = 0x08;
 
-	protected $id = self::ACTIVATOR_RAIL;
-
-	public function getName() : string{
-		return "Activator Rail";
+	protected function getConnectionsForState() : array{
+		return self::CONNECTIONS[$this->meta & ~self::FLAG_POWERED];
 	}
-
-	//TODO
 }
